Update script.js
After ten million years of inactivity, shifting interests and laziness, I have finally decided to work on some more code. Recently I have been looking into the difficult feat of coding 3d objects with the inclusion of animation so for today's update, I have integrated the Animation constructor to aid in easy-to-type animation assembly. The Animation constructor has 4 different properties, methods and arguments to it: start, stop, speed and func. Start is a method to initiate the animation process while stop will pause the animation. Entering start() again will continue the animation from where it was left off. The speed property is used to set the rate of the animation. If nothing is entered for the speed property, it will default to 0. As the speed property uses multiplication, you will need to set a value above or below 0 in order for the animation to actually take place. Also make sure the value for speed is declared before you start the animation because then the animation won't have the speed to refer to. The func argument is the callback that is used during the requestAnimationFrame loop. When putting a callback for the func argument, make sure the callback has it's own argument because the time value within requestAnimationFrame is inserted into that callback and animation will not work if an argument is not included.

Here's example code of how to operate this constructor:

'''var div = document.querySelector(".animated");

var anim = new Animation(function(time) {
  div.style.transform = "translate(" + time + "px)";
});

anim.speed = 0.5;

anim.start();'''