Update script.js
I'm trying to get a bit more active in regards to updating this js file. This time around I decided to change the Mouse object I made some time ago.
Mouse.location has been changed to Mouse.coords and both Mouse.coords and Mouse.target when not defined yet will return "unavailable"
Mouse.click() since it is meant to be based on the mouse no longer has arguments for clicking on whatever position on the DOM. This method now will only simulate a click on whatever the mouse is currently hovering over (or whatever the last known coordinate for the mouse was if the mouse is no longer detectable)
Today I saw that Object.prototype.watch is a method that is used to look for a change in value and provides a callback for when there is a change but I also noticed that this method is not compatible with most browsers so I decided to make my own (which only applies if it doesn't already work for your browser) which is formatted very similarly: You enter an object before .watch and put the property you want to monitor in the first argument. The second argument is the callback that runs when the value changes but unlike the original, the callback doesn't have 3 arguments you can put in, rather the only argument is has is an object which will return values for .value and .oldValue. These names explain themselves quite well by the fact that .value is the current value of whatever property you're looking at and .oldValue is the previous value before it. An example of how to use .watch() has been provided as a comment in the code