Update script.js
I've added some stuff that isn't super huge or important but I thought was practical:

forEach is apparently an already existing method but google couldn't seem to recognize it when I was writing code using it so I made my own method for it. forEach is nothing special but it really helps write certain for loops a lot quicker. You use forEach by plugging it into an array or a nodelist and putting in a function for the callback with one argument which represents every value in the array/nodelist.

swap is an array method simply meant to switch around two different values. I thought this would be a nice method to have after just toying around with code and wondering how I could switch around an array's values. Examples of how to use forEach and swap are provided in the code